requirements.txt
django==3.2.16
djangorestframework==3.14.0
psycopg2-binary==2.9.3
Pillow==9.3.0
djangorestframework==3.14.0
requests
celery
redis